requirements.txt
flask>=3.0
telethon>=1.40
pysocks>=1.7
requests>=2.31
pywebview>=5.0
